Used helper class for initExplorer and initPreview
